requirements.txt
# AI API App E3 - Dépendances globales
# Compétences: C9, C10, C11, C12, C13

# ═══ API REST (C9) ═══
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pyjwt==2.8.0
prometheus-client==0.20.0

# ═══ Machine Learning ═══
scikit-learn==1.4.2
xgboost==2.0.3
pandas==2.2.1
numpy==1.26.4

# ═══ MLflow (Registry + Tracking) ═══
mlflow==2.10.0
boto3==1.34.55

# ═══ Monitoring (C11) ═══
evidently==0.4.22

# ═══ Interface GUI (C10) ═══
streamlit==1.32.2
plotly==5.20.0
requests==2.31.0

# ═══ Tests (C12) ═══
pytest==8.1.1
pytest-cov==5.0.0
pytest-asyncio==0.23.6
pytest-html==4.1.1
httpx==0.27.0
ruff==0.3.0